+++ moderatorFrontend/src/jsMain/kotlin/views/guestCheckIn/guestCheckInOverview/GuestCheckInOverview.kt
@@ -6,8 +6,9 @@ import react.*
 import util.Strings
 import util.get
 import views.common.*
-import views.common.ToolbarButton
-import views.common.ToolbarViewConfig
+import views.common.ToolbarView.ToolbarButton
+import views.common.ToolbarView.ToolbarViewFc
+import views.common.ToolbarView.ToolbarViewConfig
 import views.guestCheckIn.AddGuestCheckInConfig
 import views.guestCheckIn.AddGuestCheckInFc
 import webcore.*
